refactor: move `ProcessTrackerRunner` to `diesel_models` crate
