+++ crates/scheduler/src/utils.rs
@@ -231,7 +231,7 @@ pub async fn get_batches(
 }
 
 pub fn get_process_tracker_id<'a>(
-    runner: crate::types::ProcessTrackerRunner,
+    runner: storage::ProcessTrackerRunner,
     task_name: &'a str,
     txn_id: &'a str,
     merchant_id: &'a str,
